refactor(studio): fix inverted import + derive FILTERS from schema
- Move StudioMember type to members.ts (canonical home); route re-exports it,
  team-section imports from modules — no component→route dependency
- FILTERS derived from MemberTrackSchema.options — adding a track enum value
  automatically includes it in the filter bar

Refs #77
