+++ app/modules/content/members.ts
@@ -10,6 +10,16 @@ import { createValidator } from './processor';
 export type Member = MemberFrontmatter & { slug: string };
 export type MemberRegistry = Record<string, Member>;
 
+export type StudioMember = {
+  slug: string;
+  name: string;
+  track: MemberTrack;
+  role: string;
+  bio: string;
+  avatar: string;
+  linkedin?: string;
+};
+
 const memberValidator = createValidator<MemberFrontmatter>(
   MemberFrontmatterSchema,
   'Member',
